refactor(core): split event task tools
