+++ packages/junior/src/chat/tools/update-event-task.ts
@@ -1,0 +1,114 @@
+import { z } from "zod";
+import { getDb } from "@/chat/db";
+import { saveActiveEventTask } from "@/chat/event-tasks/store";
+import {
+  changesEventTaskTrigger,
+  cleanEventTaskEvents,
+  eventTaskSuccess,
+  eventTaskToolResultSchema,
+  eventTaskTriggerSchema,
+  requireEventTaskSlackContext,
+  requireSupportedEventTaskTrigger,
+  writableEventTask,
+} from "@/chat/event-tasks/tool-support";
+import type { EventTask } from "@/chat/event-tasks/types";
+import {
+  normalizeEventIdentifier,
+  type ResourceEventCatalog,
+} from "@/chat/resource-events/catalog";
+import { zodTool } from "@/chat/tool-support/zod-tool";
+import { ToolInputError } from "@/chat/tools/execution/tool-input-error";
+import type { ToolRuntimeContext } from "@/chat/tools/types";
+
+/** Create the core tool that updates an event task in this destination. */
+export function createUpdateEventTaskTool(
+  context: ToolRuntimeContext,
+  catalog: ResourceEventCatalog,
+) {
+  return zodTool({
+    approvalMode: "review",
+    annotations: {
+      destructiveHint: true,
+      idempotentHint: true,
+      openWorldHint: true,
+      readOnlyHint: false,
+    },
+    executionMode: "sequential",
+    description:
+      "Update the instruction, registered trigger, or credential use for an event task in this Slack channel or DM. Event tasks created from other threads in the same destination are manageable here.",
+    inputSchema: z
+      .object({
+        taskId: z.string().min(1),
+        task: z.string().trim().min(1).max(4000).optional(),
+        trigger: eventTaskTriggerSchema(catalog).optional(),
+        credentialMode: z
+          .enum(["system", "creator"])
+          .describe(
+            "Set creator to make the task's original creator credentials available, or system to disable them. Creator always means the task's createdBy actor, never the current requester. Only that original creator may enable creator mode. Omit to leave unchanged.",
+          )
+          .optional(),
+      })
+      .strict(),
+    prepareArguments(args) {
+      const input = args as {
+        taskId: string;
+        task?: string;
+        trigger?: z.input<ReturnType<typeof eventTaskTriggerSchema>>;
+        credentialMode?: "creator" | "system" | null;
+      };
+      const { credentialMode, ...prepared } = input;
+      return credentialMode ? { ...prepared, credentialMode } : prepared;
+    },
+    outputSchema: eventTaskToolResultSchema,
+    async execute(input) {
+      const current = await writableEventTask(context, input.taskId);
+      if (input.trigger) {
+        requireSupportedEventTaskTrigger(catalog, input.trigger);
+      }
+      const { actor } = requireEventTaskSlackContext(context);
+      const isCreator = actor.userId === current.createdBy.slackUserId;
+      if (input.credentialMode === "creator" && !isCreator) {
+        throw new ToolInputError(
+          "Only the event task creator can enable creator credential use.",
+        );
+      }
+      if (
+        input.task === undefined &&
+        input.trigger === undefined &&
+        input.credentialMode == null
+      ) {
+        throw new ToolInputError("Event task update requires a change.");
+      }
+      const nextTrigger = input.trigger
+        ? {
+            namespace: input.trigger.namespace,
+            identifier: normalizeEventIdentifier(
+              catalog,
+              input.trigger.namespace,
+              input.trigger.identifier,
+            ),
+            resourceType: input.trigger.resourceType,
+            label: input.trigger.label,
+            events: cleanEventTaskEvents(input.trigger.events),
+          }
+        : current.trigger;
+      const changesExecution =
+        (input.task !== undefined && input.task !== current.task.text) ||
+        changesEventTaskTrigger(current.trigger, nextTrigger);
+      const next: EventTask = {
+        ...current,
+        credentialMode:
+          changesExecution && !isCreator
+            ? "system"
+            : (input.credentialMode ?? current.credentialMode),
+        task: input.task ? { text: input.task } : current.task,
+        trigger: nextTrigger,
+      };
+      const saved = await saveActiveEventTask(getDb(), next);
+      if (!saved) {
+        throw new ToolInputError("Event task is no longer active.");
+      }
+      return eventTaskSuccess(saved, catalog);
+    },
+  });
+}
